+++ include/Servicio.h
@@ -2,6 +2,7 @@
 #define SERVICIO_H
 
 #include <string>
+#include "Funciones.h"
 using namespace std;
 
 class Servicio
@@ -9,19 +10,19 @@ class Servicio
     private:
         int codigo;
         char nombre[50];
-        float valor;
+        double valor;
         int codCategoria;
         bool estado;
 
     public:
         int getCodigo();
         string getNombre();
-        float getValor();
+        double getValor();
         int getCodCategoria();
         bool getEstado();
         void setCodigo(int _codigo);
         void setNombre(string _nombre);
-        void setValor(float _valor);
+        void setValor(double _valor);
         void setCodCategoria(int _codCateg);
         void setEstado(bool _estado);
